[textanalytics] Add DebugView for ExtractKeyPhraseResultCollection as an inner class
