+++ sdk/textanalytics/Azure.AI.TextAnalytics/src/ExtractKeyPhrasesResultCollection.cs
@@ -4,6 +4,7 @@
 using System.Collections.Generic;
 using System.Collections.ObjectModel;
 using System.Diagnostics;
+using System.Linq;
 
 namespace Azure.AI.TextAnalytics
 {
@@ -32,5 +33,43 @@ namespace Azure.AI.TextAnalytics
         /// on this batch of documents.
         /// </summary>
         public string ModelVersion { get; }
+
+        /// <summary>
+        /// Debugger Proxy class for <see cref="ExtractKeyPhrasesResultCollection"/>.
+        /// </summary>
+        internal class ExtractKeyPhrasesResultCollectionDebugView
+        {
+            private ExtractKeyPhrasesResultCollection BaseCollection { get; }
+
+            public ExtractKeyPhrasesResultCollectionDebugView(ExtractKeyPhrasesResultCollection collection)
+            {
+                BaseCollection = collection;
+            }
+
+            [DebuggerBrowsable(DebuggerBrowsableState.RootHidden)]
+            public List<ExtractKeyPhrasesResult> Items
+            {
+                get
+                {
+                    return BaseCollection.ToList();
+                }
+            }
+
+            public TextDocumentBatchStatistics Statistics
+            {
+                get
+                {
+                    return BaseCollection.Statistics;
+                }
+            }
+
+            public string ModelVersion
+            {
+                get
+                {
+                    return BaseCollection.ModelVersion;
+                }
+            }
+        }
     }
 }
